+++ src/SignatureReader.h
@@ -21,7 +21,7 @@ private:
     {
         std::vector<std::uint8_t> m_buffer; // Byte values to compare.
         std::string               m_type;   // Name of the file type.
-        std::set<std::size_t>     m_skip;   // A set of byte indexes in the query whose value should not be taken into account.
+        std::set<std::uintmax_t>  m_skip;   // A set of byte indexes in the query whose value should not be taken into account.
                                             // For example, the signature for a WAV file is: 0x52 0x49 0x46 0x46 ?? ?? ?? ?? 0x57 0x41 0x56 0x45
                                             // The fields of the corresponding Signature object should be:
                                             // m_buffer: 0x52 0x49 0x46 0x46 0x57 0x41 0x56 0x45
