+++ test/IOHandlerMock.cc
@@ -20,7 +20,7 @@ void IOHandlerMock::close()
 {
 }
 
-bool IOHandlerMock::read(std::uint8_t* o_buffer, std::size_t buffer_size)
+bool IOHandlerMock::read(std::uint8_t* o_buffer, std::uintmax_t buffer_size)
 {
     if (!o_buffer
         || buffer_size == 0
@@ -34,7 +34,7 @@ bool IOHandlerMock::read(std::uint8_t* o_buffer, std::size_t buffer_size)
     return !m_io_fail;
 }
 
-bool IOHandlerMock::write(const std::uint8_t* i_buffer, std::size_t buffer_size)
+bool IOHandlerMock::write(const std::uint8_t* i_buffer, std::uintmax_t buffer_size)
 {
     if (!i_buffer
         || buffer_size == 0
@@ -46,7 +46,7 @@ bool IOHandlerMock::write(const std::uint8_t* i_buffer, std::size_t buffer_size)
     return !m_io_fail;
 }
 
-bool IOHandlerMock::seek(std::uint64_t offset)
+bool IOHandlerMock::seek(std::uintmax_t offset)
 {
     m_id = offset / Hexit::ChunkCache::capacity;
     return !m_io_fail;
@@ -54,7 +54,7 @@ bool IOHandlerMock::seek(std::uint64_t offset)
 
 std::uint8_t* IOHandlerMock::data() { return reinterpret_cast<std::uint8_t*>(m_data); }
 
-std::uint64_t IOHandlerMock::load_count() const { return m_load_count; }
+std::uintmax_t IOHandlerMock::load_count() const { return m_load_count; }
 
 void IOHandlerMock::mock_io_fail(bool should_fail)
 {
